Making the playbook modular.
